Refactors route handling into modular structure
Moves route definitions to a separate file for improved modularity and maintainability. Simplifies the main request handler by delegating route logic to the new module. Updates route functions to take request, response, and decoded user data as parameters.
